Prevent registerProtocolHandler from nagging the user after they say no.
Currently if a user visits a site that calls registerProtocolHandler(), and
they deny that sites request, then the next time they visit the site it will
ask them again. This patch changes the semantics of denying a request from
doing nothing to remembering the site, but not registering it as the default.
Handlers retained in this way can still be removed from the settings screen.

TEST=Unit tests provided.

Review URL: http://codereview.chromium.org/6976018

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@86880 4ff67af0-8c30-449e-8e8b-ad334ec8d88c