Retain support for 'UnlimitedStorage' in the QuotaManager.
Track unlimited_usage in parallel with total global_usage. The latter value is the sum of all usage by any origin (same as before). The new value is the amount of the total that is attributable to origins with the unlimited storage permission.

Widened the GlobalUsageCallback to take two int64 params, one for the total global_usage and the other for the unlimited_usage.

Modified the eviction logic to exempt origins with the unlimited storage right from eviction.

Modified the behavior of QuotaManager::GetUsageAndQuota(origin) method to return kint64max for unlimited origins. Also modified the return value in the limited case to be 10% of the total limit for temp storage, unless the system is over the total limit in which case they're capped at their current usage level.

  // Allow an individual host to utilize no more than 10% of the total 
  // pool available for temp storage.
  int64 host_quota = quota() / 10;

  // If total temp usage is over-budget, stop letting new data in until we reclaim space.
  limited_global_usage = global_usage() - unlimted_global_usage();
  if (limited_global_usage > quota())
    host_quota = std::min(host_quota, host_usage());  
  return host_quota;

At this time, the total limit for temp storage is 1G, that gives each site 100M to work with (no prompts involved).

BUG=61676
TEST=unittests
Review URL: http://codereview.chromium.org/6962005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@87207 4ff67af0-8c30-449e-8e8b-ad334ec8d88c