Register handlers with operating system when there is a new default handler.
Previously registration happened when a protocol became known to the registry for
the first time. This has two problems:
1. most seriously, if the user says no to register a protocol it would still be
registered with the OS
2. less seriously, if another application had "registered over the top" of Chrome
and then the user registered a handler in Chrome again, it would not register
with the OS.

Now we register with the OS whenever a protocol is made the default, iff there
wasn't a default for that protocol before.

BUG=83556
TEST=Unit tests added.


Review URL: http://codereview.chromium.org/7096003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@87362 4ff67af0-8c30-449e-8e8b-ad334ec8d88c