Fix enable / disable handlers radio buttons in the content settings page.
Currently this crashes chrome in debug mode. The problem is that a selector
used to decorate radio buttons is too broad and attaches a listener to radio
buttons that later call callbacks with data they aren't expecting.

This change fixes this problem by making the selector in question more
specific.

TEST=Click enable / disable for the handlers section in chrome://settings/content and verify there's no crash.


Review URL: http://codereview.chromium.org/7096008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@87365 4ff67af0-8c30-449e-8e8b-ad334ec8d88c