Prevent model association from completing if passphrase is not available.
Without this fix password model associator will continue association as long as there is a default nigori, even if it does not have the latest decryption key. The reason it worked was syncer would not give us the nodes that we cant decrypt instead they would just be labeled as conflicts.
So DTM would tell PSS send sync_configure_done and we would dismiss the UI.

With this fix we would not send sync_configure_done unless we have the latest passphrase. 
Note: with this fix without the passphrase sync will simply stop(and show error). before this fix sync MIGHT continue.(behavior without the fix: if this is a new profile then sync will stop completely. if the profile had already been setup and the passphrase changed in another client then all unencrypted types will keep syncing in this client).

Now we are unifying them to NOT sync. However if there is a reason to let them sync then we need to file a bug and think of ways in which we could send sync configure done but still have undecrypted items.
BUG=82496
TEST=


Review URL: http://codereview.chromium.org/6966027

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@87411 4ff67af0-8c30-449e-8e8b-ad334ec8d88c