Replace ButtonDropDown menu implementation (Menu2) with MenuItemView.
- Added views::MenuDelegate::WillHideMenu()
- Added EMPTY to views::MenuItemView::Type to suppress delegate call.
- Minor fixes for views::MenuModelAdapter

There was a Windows-only views unit test that a click-drag would post the ButtonDropDown menu.  This test relied on TrackPopupMenu not blocking for input if the menu was empty.  Because MenuItemView does block, this test could not work as a unit test.  I removed this test and added an interactive ui test to do the same thing.

BUG=chromium-os:13887
TEST=none


Review URL: http://codereview.chromium.org/7044016

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@87490 4ff67af0-8c30-449e-8e8b-ad334ec8d88c