ntp4: Add a new page when the user starts a drag
(the page is deleted if unused)

also, change the dot label from a contenteditable span to an input with -webkit-appearance: caret. This is necessary because the span is not really editable when there's no content (e.g., if the user backspaces all the way to 0 characters, the focus evaporates). IMO this is a bug, either in webkit or in the spec. (see https://bugs.webkit.org/show_bug.cgi?id=20288 , which has been ignored for the past 2 1/2 years)

as a side effect of how pages are persisted, empty pages get deleted when the page is next loaded. I think this kind of makes sense.

BUG=none
TEST=manual

Review URL: http://codereview.chromium.org/7109001

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@87727 4ff67af0-8c30-449e-8e8b-ad334ec8d88c