Now makes HistoryMenuBridge more multi-profile friendly.
HistoryMenuBridge was listening for HISTORY_LOADED notifications from all
sources, which meant that if you are running with multiple profiles but the
non-default profile's history DB loaded *first*, it would call CreateMenu()
with a null HistoryService.

This patch makes us only listen for HISTORY_LOADED for the default profile, and
also adds a DCHECK to catch this in the future.

BUG=84559
TEST=Run sync_integration_tests (which use multiple profiles) on the build bots.

Review URL: http://codereview.chromium.org/7112005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@87750 4ff67af0-8c30-449e-8e8b-ad334ec8d88c