Make WebUILogin work on non-TOUCH_UI builds.
This CL splits the components that are TOUCH_UI specific out of the WebUILogin and into their own seperate class. There is also a bunch of little bolierplatey type things that need to be moved out of TOUCH_UI defines all over the place. The WebUILoginDisplay is been modified to support both code paths via a few simple #if defined guards. Obviously builds will still need for OS_CHROMEOS to be defined for the login screen to work.

The TOUCH_UI screen works as expected at the moment expect for a grey border on the bottom and right sides. This appears to be an issue with the implementation of the keyboard not the login screen. This issue will probably not get resolved since this keyboard is an interim solution.

Currently the SAV on the non-TOUCH_UI builds does not appear though it does exist. This is due to the TabContents of the DOMView being in its own widget and occulding it. I believe the work being done by MSK to move the SAV into its own overlay widget will fix this.

Patch submitted by Ryan Harrison <rharrison@chromium.org>

BUG=chromium:15926
TEST=manual

Review URL: http://codereview.chromium.org/7020037
Patch from Ryan Harrison <rharrison@chromium.org>.

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@87822 4ff67af0-8c30-449e-8e8b-ad334ec8d88c