Partially deprecate the old webstorePrivate.beginInstall method
We are still allowing it to be used to do installs, but with this change it
will no longer prevent the confirmation dialog from appearing. Once the web
store completely transitions to use the newer beginInstallWithManifest2
method, we can remove beginInstall alltogether.

BUG=75821
TEST=Installing an app/extension from the web store should cause an install
confirmation prompt to appear after the item downloads.

Review URL: http://codereview.chromium.org/7131001

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@88204 4ff67af0-8c30-449e-8e8b-ad334ec8d88c