Fix crash when you uninstall a permissions-upgrade disabled extension.
When an extension is disabled because a newer version requires higher
permissions, if you then uninstall the extension while the infobar warning
is still up, you'll get a crash if you switch to the tab that had the infobar.
This patch makes the infobar notice the extension uninstall and remove itself.

BUG=84826
TEST=As detailed in description above (and outlined in bug report too).


Review URL: http://codereview.chromium.org/7003068

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@88458 4ff67af0-8c30-449e-8e8b-ad334ec8d88c