Fix crash when swapping in a prerendered page for a page which has an unload handler.
***
Note: this is the fix without a browser test, for expediency. A CL with the test is in http://codereview.chromium.org/7046053/
***

The root cause PrerenderManager::OnCloseTabContentsDeleter was never unregistering itself as a TabContentsDelegate in CloseContents. 

r87606 exposed this by changing the TabContents destructor to deference the delegate if it's non-NULL. 

BUG=85350
TEST=None

Review URL: http://codereview.chromium.org/7108032

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@88478 4ff67af0-8c30-449e-8e8b-ad334ec8d88c