Retry r88137:
Add code to calculate the dominant color for a favicon.

Currently we calculate the dominant/representative color only for those favicons we need it for (i.e. the ones shown on the NTP). We don't do any caching either in memory or in the favicon db but that can be tacked on later if deemed suitable.

Code in color_analysis.* authored by dtrainor

BUG=none
TEST=trybots

Review URL: http://codereview.chromium.org/7031078

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@88600 4ff67af0-8c30-449e-8e8b-ad334ec8d88c