Bugfix for tabstrip painting with tall tabstrip/short theme_frame image.
If the tabstrip is higher than the theme_frame pixmap height, we wouldn't paint enough. As a fix, I make sure we cover the to the bottom of the tabstrip at least. As a safety measure (to avoid regressions) I make sure we paint at least to the bottom of the theme_frame height if that was more.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/7134027

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@88732 4ff67af0-8c30-449e-8e8b-ad334ec8d88c