CrOS - Fix problem attaching files in Gmail and Hotmail
This is a workaround for a race condition when closing the file browser dialog on Chrome OS - sometimes the window closes before the main UI thread gets the message that a file was selected.  I'm working on a more definitive solution, but I needed something small enough to backport to R13.  Also added some LOG lines when the dialog is opened and closed to help troubleshoot related issue.

BUG=chromium-os:16429,chromium-os:16267
TEST=manual, see bugs.  Basically, ensure you can attach at least 3 files in several web-based email providers.


Review URL: http://codereview.chromium.org/7146020

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@89171 4ff67af0-8c30-449e-8e8b-ad334ec8d88c