Linux: remove a bunch of code from gnome keyring integration that was needed to work around bugs in old versions of the gnome keyring client library that we no longer support.
Ultimately removing this may allow us to no longer force the keyring to unlock
right away, since all remaining code uses APIs that have that as a side effect.

If that pans out, and if gnome keyring is smart enough to not prompt to unlock
unless it actually needs to return secrets, then this change may lead to some
improvement for bug 85285 by not displaying the unlock prompt unless a password
would actually be filled in when visiting a site with a login form.
Review URL: http://codereview.chromium.org/7155011

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@89218 4ff67af0-8c30-449e-8e8b-ad334ec8d88c