Add custom NSView as a titlebar of a Panel so we can custom-paint it.
The XIB file change is adding a view and linking it with PanelTitlebarViewCocoa class. The view is not parented to the Panel window since XIB can't set a custom view to shadow the regular titlebar - so we hook it up in awakeFromNib for the panel.

Review URL: http://codereview.chromium.org/7108069

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@89564 4ff67af0-8c30-449e-8e8b-ad334ec8d88c