When a user logs into sync, the appropriate cookies are retrieved so that she is already logged into Google web services, and does not need to enter her username and password again.
This feature is on by default, but can be turned off by specifying
--disable-auto-login on the command line or the about:flags page.

BUG=None
TEST=Make sure the browser has no google or youtube cookies.  Either clear
all the cookies or start with a brand new profile.  Go to menu item
"Wrench / Options", go to the tab "Personal stuff", and click the "Enable
these features..." button to enable sync.  Follow the wizard to login to your
google account and finish the sync process.  Once terminated, browse to
gmail.com and you should be already logged in.

Review URL: http://codereview.chromium.org/7121014

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@89842 4ff67af0-8c30-449e-8e8b-ad334ec8d88c