Revert 90511 - Use pixmaps and EGLImages to transport image data between GPU process and browser. Behind a compile TOUCH_UI flag.
It uses the MAC OSX AcceleratedSurface IPCs to communicate between the GPU process and browser. The major difference between the OSX display path is that I send an ACK back after AcceleratedSurfaceSetIOSurface because the process of binding a texture to the pixmap may destroy the contents of the pixmap.

BUG=none
TEST=by hand on Linux (w/ and w/o TOUCH_UI), Windows, and Mac

Review URL: http://codereview.chromium.org/6987014

TBR=backer@chromium.org
Review URL: http://codereview.chromium.org/7261016

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@90512 4ff67af0-8c30-449e-8e8b-ad334ec8d88c