Store profile avatar to preferences
This change adds a preference to store avatar icons.

For this change I extended the idea of kProfileAvatarIconMap pref into a full blown class. The class is responsible for mantaining a list of sorted profiles that can be displayed without having to load profiles from disk.

Once this is chcked in there are several TODOs:
  - remove the ProfileManager::GetNumberOfProfiles(), etc... functions that I added earlier
  - add UI to customize the avatar icon
  - create a single instance of the new ProfileInfoCache object some where. Maybe from ProfileManager?

BUG=
TEST=


Review URL: http://codereview.chromium.org/7155015

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@90622 4ff67af0-8c30-449e-8e8b-ad334ec8d88c