[Mac] Don't crash when double-clicking a LinkInfoBar link rapidly.
The issue is that the InfoBarController gets deallocated when the animation
finishes running. But, the second click event can be delievered either before or
after that happens, meaning that sometimes a zombie can be messaged. Use the
performSelector:withObject:afterDelay: hack-and-slash to "work around" this.

BUG=87201
TEST=Enter a string of a known hostname (like crash) and get "Did you mean http://crash?" on the search results page. Double click the link quickly and don't crash. Try multiple times for safety.


Review URL: http://codereview.chromium.org/7277008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@90678 4ff67af0-8c30-449e-8e8b-ad334ec8d88c