Fix Chrome About dialog unicode 'official build' label to display properly.
We recently added a UTF8 encoded 'official build' localized string to the About dialog, but elsewhere in the file that string is treated as containing ASCII. This causes garbage to be displayed in ie. Russian locale.

This affects Chromium builds only, since this is not done for Google Chrome builds.

BUG=87370
TEST=Open Chromium in Russian locale. No garbage should been seen in the dialog (see screenshot in bug).
Review URL: http://codereview.chromium.org/7274001

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@90776 4ff67af0-8c30-449e-8e8b-ad334ec8d88c