Autocomplete: fix array size to make size check work
AutocompleteMatch has two assertions that are asserting some arrays
are the same size as an enum.  But the arrays are initialized with
a static size, which is taken from the enum, so the checks always
pass.  Removing the static size restores the spirit of the assertion.

While I'm at it, change the DCHECK into a compile-time check.

Review URL: http://codereview.chromium.org/7274009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@90848 4ff67af0-8c30-449e-8e8b-ad334ec8d88c