Revert 89600 - Treat ERR_CONNECTION_CLOSED as a succesful download.
=========================
Note: This revert is only temporary, to collect statistics for how frequently
this case is encountered in the wild. This will be re-reverted shortly.
=========================

Some servers will send a Content-Length which exceeds the number of bytes sent
in the message body.

Although this violates the RFC spec, other major browsers permit it
and treat a download as complete in that case. Chrome did as well
prior to M12, and now some users are unable to access the item that
they downloaded.

Attempt #2.

BUG=85549
TEST=Go to server which advertises larger Content-Length than number of bytes
sent, confirm that the download is no longer reported as interrupted.


Review URL: http://codereview.chromium.org/7237031

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@91769 4ff67af0-8c30-449e-8e8b-ad334ec8d88c