Made Print Preview tests autogenerated from print_preview.js.
Now that the generator exists, make the print preview tests use it as well.  Since the test was flaky when I found it, porting FLAKY_ to the javascript test name.

Since Print Preview needs the preloading, provided a way to tell the generator to not AddLibrary.

BUG=82437
R=thestig@chromium.org,dtseng@chromium.org
TEST=browser_tests --gtest_filter=PrintPreview*.*


Review URL: http://codereview.chromium.org/7304014

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@91787 4ff67af0-8c30-449e-8e8b-ad334ec8d88c