Make header copyable in views info bubble; add a comment about description.
This change makes the info bubble text copyable by changing it from a Label
to a Textfield (similar to how it's done in about_chrome_view.cc for the 
version text). Unfortunately, the same can't yet be done to the description
label since currently a textfield cannot be multiline. I've added a comment
to that extent.

BUG=52526
TEST=manual

Review URL: http://codereview.chromium.org/7316022

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@92169 4ff67af0-8c30-449e-8e8b-ad334ec8d88c