Fix leaking request IDs in webRequest API
The current webRequest API did not guarantee that either onCompleted or onErrorOccurred is called for each request. If an extension fills a REQUEST_ID -> URL hashmap for all requests at the onBeforeRequest event, and tried to delete the entries on onCompleted and onErrorOccurred, some entries remained in the hashmap.

BUG=86139
TEST=execute example extension attached to bug

Review URL: http://codereview.chromium.org/7190007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@92331 4ff67af0-8c30-449e-8e8b-ad334ec8d88c