Add WebUI certificate viewer to eventually replace GTK certificate viewer.
Hidden behind the gyp flag webui_certificate_viewer, this dialog will replace the GTK certificate viewer as part of the nogtk effort. This CL only adds the new certificate viewer but does not use it unless compiled with webui_certificate_viewer=1.

This is the first stage, which adds the viewer and only shows the "General" certificate details (As shown in the General tab in the GTK viewer). The next steps are to show the detailed information, resolve some minor display and functionality issues on desktop builds and then make this the standard certificate viewer.

BUG=None
TEST=Manually viewed certificate details for several secure sites.


Review URL: http://codereview.chromium.org/7331016

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@92541 4ff67af0-8c30-449e-8e8b-ad334ec8d88c