chromiumos: 3G: Keep track of a cellular provider's APN list.
Added provider_apn_list() method to NetworkDevice. Returns
a list of the APNs associated with the home provider. Each
APN may include a non-localized name, a localized name,
and, the two-letter language code of the localized name (if
the localized name exists).

This is part 1 of implementing the feature described in
http://crosbug.com/14290.

BUG=chromium-os:14290
TEST=Verified with log statements that the contents of the
flimflam Cellular.APNList device property are correctly
populating the APN list in NetworkDevice.

R=nkostylev@chromium.org

Review URL: http://codereview.chromium.org/7342051

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@92565 4ff67af0-8c30-449e-8e8b-ad334ec8d88c