[Mac] Properly fix cursor restore after closing context menu.
My previous change (http://crrev.com/92370) missed the case where the
previous cursor sent by the renderer is the same as the one that should be
displayed after the context menu is closed. This case is special because the
renderer will not re-send the cursor if there was no change.

To handle it, first restore to the previous cursor before sending a mouse
move event to the renderer.

BUG=64436
TEST=manual

Review URL: http://codereview.chromium.org/7400004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@92775 4ff67af0-8c30-449e-8e8b-ad334ec8d88c