chromiumos: Notify NetworkManager observers when the service list becomes empty.
In UpdateNetworkServiceList, if the service list was empty,
NetworkManager observers were not being notified. Normally,
when there was at least one service, RequestNetworkServiceInfo
would be called, and the callback passed in that call would
end up calling NotifyNetworkManager. But with no services,
there was no path that would end up calling NotifyNetworkManager.
This change adds such a call for the specific case of an empty
service list.

BUG=none
TEST=Open the internet settings page. Disable Wi-Fi and enable
mobile data. Have a SIM inserted that allows the device to
register on a cellular network, so that there is exactly one
network showing in the "Wireless networks" section. Make sure
that PIN-locking is enabled for the SIM, but that it is currently
unlocked. Now try to change the PIN, but enter an incorrect PIN
three times, causing the SIM to become PUK-locked (I did this
step on the command line using the "pin" script in
/usr/local/lib/flimflam/test, which makes it a bit easier to see
what's going on). When the SIM becomes PUK-locked, the device is
deregistered from the network. Running list-services on the
command line should show no services. Prior to the present
fix, the "Wireless networks" section of the settings page would
continue to show the network that had been registered, unless
the settings page was refreshed.  With the present fix, the
network should disappear from the display, leaving no wireless
networks.

R=stevenjb@chromium.org,nkostylev@chromium.org,dpolukhin@chromium.org

Review URL: http://codereview.chromium.org/7400031

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@93046 4ff67af0-8c30-449e-8e8b-ad334ec8d88c