bookmarks: Don't expose an implementation detail function through BookmarkNode public API.
This function is only interesting for BookmarkModel and at this moment is just used there.
Instead of exposing it through the BookmarkNode class just do it internally in BookmarkModel.
This has the advantage of preventing further external clients of using it.
This is a responsibility of BookmarkModel, and just it knows when to call this function.
The consumers of BookmarkNode API don't know and shouldn't know where/when it's necessary to call
InvalidateIcon(), so lets the model do the work for them.

BUG=None
TEST=None

R=sky@chromium.org

Review URL: http://codereview.chromium.org/7401013

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@93306 4ff67af0-8c30-449e-8e8b-ad334ec8d88c