[Sync] Support receiving early passphrase required information in Sync UI.
We now always send the configure command, even when it's just a passphrase
being entered. This works because the datatypemanager now is smart enough
to not reconfigure if the datatypes are the same and we were already in a
configured state. The configure command now also supports using the explicit
passphrase prompted when one is already required.

BUG=87797,87572
TEST=First time sync on an account that already had an explicit passphrase. Also all other passphrase prompt scenarios.


Review URL: http://codereview.chromium.org/7466024

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@93549 4ff67af0-8c30-449e-8e8b-ad334ec8d88c