The problem:
When the user has the "Clear cookies and other site and plug-in data when I close my browser" setting enabled, all appcache data is deleted when the user closes the browser. The intention is to protect an appcache from deletion, if the user has installed a hosted app which refers to the origin of the appcache.

BUG=79604
TEST=ChromeAppCacheService.KeepOnDestruction, ChromeAppCacheService.RemoveOnDestruction


Review URL: http://codereview.chromium.org/7210006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@93597 4ff67af0-8c30-449e-8e8b-ad334ec8d88c