Add additional check for required passphrase in SetupSync
The sync integration tests use a method called SetupSync to initialize a
sync client. The step that performs an initial sync can sometimes
remain incomplete due to a required passphrase.

r93873 relocated the check for SET_PASSPHRASE_FAILED in SetupSync. It
turns out that this check needs to be made before and after initial
sync, since SET_PASSPHRASE_FAILED can be triggered in a couple of
places.

This patch adds a check at the original place, and should fix the flaky
test SetPassphraseAndThenSetupSync.

TBR=zea@chromium.org
BUG=90371, 90397
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/7477006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@94076 4ff67af0-8c30-449e-8e8b-ad334ec8d88c