Fixed slider thumb in the mediaplayer progress.
Now "-webkit-appearance: sliderthumb-*" is used for a shadow element in input-range. The input-range element itself implicitly has "-webkit-appearance: slider-*" style that causes it to be rendered as a slider bar. Changing its style makes it transparent (we have separate DIV to show it in a special way.

BUG=chromium-os:17030
TEST=Check that media player progress looks and behaves properly. Thumb is graggable, click the bar puts the thumb there.


Review URL: http://codereview.chromium.org/7488055

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@94081 4ff67af0-8c30-449e-8e8b-ad334ec8d88c