Re-enable TaskManagerBrowserTest.ReloadExtension
In the past this test had been crashy, and I was able to reproduce the crash
occasionally in vmware and on the try bots. I'm no longer able to repro, so
I'm speculatively re-enabling it to see if the underlying cause is actually
fixed.

BUG=42315
TEST=none

TBR=mpcomplete@chromium.org

Review URL: http://codereview.chromium.org/7537006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@94761 4ff67af0-8c30-449e-8e8b-ad334ec8d88c