This is only the first part of the refactoring. The old method was preserved and not all call points were changed.
The second part will deal with factoring out the old OpenURL method.

Overall purpose: speed up calls, make it more flexible to plumb parameters to
the call chain, reduce code repetition.

It is necessary to add one more parameter (namely, override_encoding) to
OpenURL in order to be able to restore the page encoding upon loading the
contents. For solving this issue, it is necessary for the encoding to be
restored from the BookmarkModel.

BUG=2926

Review URL: http://codereview.chromium.org/7388007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@95013 4ff67af0-8c30-449e-8e8b-ad334ec8d88c