Alwas set saved maximized state for panels to 'false'.
For reasonss that are not entirely clear, Browser::GetSavedMaximizedState() is sometimes setting maximized to true for panels created by an extension on ChromeOS.
Since panels have no concept of a maximized state, fix this by always returning 'false'.

This also includes a patch to widget.cc that is not necessary to fix this behavior, but it fixes the case where ShowInactive gets called when the widget is maximized (which shouldn't normally happen).

BUG=chromium-os:18534
TEST=See issue + test opening panels from an extension

Review URL: http://codereview.chromium.org/7551002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@95141 4ff67af0-8c30-449e-8e8b-ad334ec8d88c