chromeos: Try to avoid LoginUtilsImpl delegate segfault.
This makes LoginUtilsImpl::OnProfileCreated() check that its
delegate is non-NULL before using it.  I'm hoping that this
will avoid a segfault that's apparently showing up during
testing when Chrome is killed during initialization.

Note that this is just a temporary workaround; the
underlying code is possibly still incorrect (or relying on
incorrect assumptions about the order in which things will
happen).

BUG=chromium-os:18269
TEST=none
TBR=altimofeev@chromium.org

Review URL: http://codereview.chromium.org/7590026

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@96208 4ff67af0-8c30-449e-8e8b-ad334ec8d88c