Limit size of typed url node.
Now, we cap the # visits for each node to 100 items, to avoid hitting the 10K
node size limit on the server. We also ignore all reload visits, since those
are not used by the omnibox suggestion algorithm and tend to be the bulk of
the visits in the large nodes we've encountered.

BUG=89460
TEST=Run unit tests.


Review URL: http://codereview.chromium.org/7550062

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@96287 4ff67af0-8c30-449e-8e8b-ad334ec8d88c