[Mac] Flip the flag for multi-profiles to on by default.
This also changes ProfileManager::GetProfileInfoCache() to be GetProfileInfo()
and returning the ProfileInfoInterface, making it easier to mock out for testing.
In the few places that require the full ProfileInfoCache, GetMutableProfileInfo()
was added.

Because a handful of Mac unittests make calls into g_browser_process for the
ProfileManager, BrowserTestHelper was expanded to set that up with a
TestingProfile and a FakeProfileInfo so that tests continue to pass.

BUG=60105
TEST=Multi-profiles is on by default.

Review URL: http://codereview.chromium.org/7648037

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@97049 4ff67af0-8c30-449e-8e8b-ad334ec8d88c