content: Move render_widget_host_view_gtk to content/ try 2.
This was r97750, which was reverted at r97756. This patch, in addition to all
the other things it did, marks OwnedWidgetGtk for export and moves it to the
ui:: namespace.

Original list of things this patch does:
- OwnedWidgetGtk now goes in ui/base/gtk/
- TruncateString moved from l10n_util:: to ui::
- GtkIMContextWrapper has part of its code split into chrome/ (IDC using code goes in RenderViewContextMenu) and the rest go in content/ (gtk using code goes with GtkIMContextWrapper).
- gtk_key_bindings_handler[_unittest] now goes in content, as it's a utility class to RenderWidgetHostGtk.

BUG=93804
TEST=existing unit tests
Original Review URL: http://codereview.chromium.org/7669040

Review URL: http://codereview.chromium.org/7708021

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@97889 4ff67af0-8c30-449e-8e8b-ad334ec8d88c