Allow sync integration tests to operate on multiple datatypes: Passwords
The sync integration tests currently use a class hierarchy where the test classes for each datatype are subclasses of LiveSyncTest. While this design worked in the past, it allows tests to work with only one datatype at a time, and therefore doesn't allow us to test the interplay between datatypes.

This patch is another in the series of patches that will move away from an inheritance model to one where test cases can operate on more than one datatype. It updates the Passwords datatype to the new model, and contains the following changes:

- LivePasswordsSyncTest is no longer a class that inherits from LiveSyncTest, but is now a namespace called passwords_helper, and contains a bunch of methods that perform various operations related to Sessions.
- SingleClientLivePasswordsSyncTest is renamed to SingleClientPasswordsSyncTest, and is a subclass of LiveSyncTest. It uses the methods in namespace passwords_helper by including its header file and attaching itself to the helper.

BUG=88510
TEST=sync_integration_tests, sync_performance_tests

Review URL: http://codereview.chromium.org/7647011

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@97936 4ff67af0-8c30-449e-8e8b-ad334ec8d88c