[Sync] Gracefully handle writing to a node when the cryptographer is not ready.
Due to allowing users to configure without a passphrase if they don't have
passwords enabled, we can get into a situation where we require encryption of
datatypes, but don't have the cryptographer enabled. When in this situation
we need to ensure we don't corrupt any data. Once the user sets their
passphrase, we will reencrypt everything, which will overwrite unsynced entries
with their encrypted versions, allowing the sync cycle to finish.

BUG=93100
TEST=sync_unit_tests --gtest_filter="*GetCommitIdsFilterEntries*"
Also to manually test: set up sync on a machine with an explicit passphrase. Set up sync without passwords enabled on a separate machine and don't enter the passphrase. Enable encryption on the first client. Attempt to add a bookmark on the second client.

Review URL: http://codereview.chromium.org/7795002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@98758 4ff67af0-8c30-449e-8e8b-ad334ec8d88c