Fix localized "platform" message for non-8bit languages
Now we use localized string for "platform", we should not
assume the text is encoded in ASCII.

To test this, manually change the UI to a non-latin language
(e.g. Japanese) and see if any irregular font is not displayed.

Webui login screen is also affected.

BUG=chromium-os:19822
TEST=manually

Review URL: http://codereview.chromium.org/7791020

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@98911 4ff67af0-8c30-449e-8e8b-ad334ec8d88c