Parse and show CWS data (reviews, number of users) in the inline install dialog.
Data is put into a ExtensionInstallUI::Prompt struct that is passed to
ShowExtensionInstallDialog implementations (currently only the GTK one actually
displays it). Ratings are shown as stars (resources were r99027). The dialog
also includes a CWS link for the item (this necessitated switching the GTK
implementation to a class so that the click handler callback could get at the
delegate and extension).

R=asargent@chromium.org

Review URL: http://codereview.chromium.org/7819006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@99205 4ff67af0-8c30-449e-8e8b-ad334ec8d88c