Fix a crash on Linux which occurs during drag drop operations in the renderer while dereferncing a NULL TabContentsWrapper pointer.
Fix is to initialize the TCW member in the DragMotion event in the WebDragDestGtk object. This function is called
at the start of the D&D operation.

Added DCHECKs wherever we deref this member. There is no need to DCHECK the bookmark helper pointer
as a TCW always has a bookmark helper object.

BUG=89388
Review URL: http://codereview.chromium.org/7830022

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@99253 4ff67af0-8c30-449e-8e8b-ad334ec8d88c