Add local histogram to see what UMA pageloads are
This adds a histogram that gets a sample each time the
UMA pageload counter is incremented.  The result is that
any time you refresh:
about:histograms/Chrome.UmaPageloadCounter
You'll see how many pageloads have taken place during
the session.  Be warned: Each refresh of that histogram
page will also add 1 to the tally ;-).

If you want to use this patch in a personal build, be
sure to modify browser_main.cc so that metrics->Start()
is called (even if this is not an official build).
Review URL: http://codereview.chromium.org/7792070

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@99289 4ff67af0-8c30-449e-8e8b-ad334ec8d88c