Revert 99295 - Add some more temporary instrumentation for diagnosing bug 93747.
The goal is to conclusively tell whether the object is freed or not when running TabStripModel::GetWrapperIndex. Also surrounded the vector that we are seeing as corrupt with some special padding bytes (64 bytes total), to see if that area of memory is getting clobbered.

BUG=93747

Review URL: http://codereview.chromium.org/7827024

TBR=eroman@chromium.org
Review URL: http://codereview.chromium.org/7831031

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@99307 4ff67af0-8c30-449e-8e8b-ad334ec8d88c