CrOS - Fix file picker to route callbacks using TabContents
Previously, when the user clicked open/save/cancel in the CrOS file picker, it would route that callback data to the frontmost tab.  Sometimes other tabs can switch to frontmost, however.  Use set_associated_tab_contents() to inform the extension function implementions of the tab to which to route the callbacks.  This also adds a test from tbarzic for the referenced bug.

BUG=chromium-os:18956
TEST=browser_tests FileManagerDialogTest.*, also verify you can open the file picker and open/save/cancel under normal conditions


Review URL: http://codereview.chromium.org/7830024

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@99327 4ff67af0-8c30-449e-8e8b-ad334ec8d88c