Remove ContentSettingsRule and ContentSettingsRules types in content_settings::PolicyProvider.
The change removes one step in the process of loading managed (= by policy) content settings form the preferences that store the policy values. Instead of loading the content settings into a vector and passing this vector to another method that stores the content settings in the in-memory map , content settings are directly stored in the in-memory map. This eliminates one pass over all content settings. That is not needed.

BUG=63656
TEST=content_settings_policy_provider_unittests.cc

Review URL: http://codereview.chromium.org/7811019

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@99632 4ff67af0-8c30-449e-8e8b-ad334ec8d88c