Implement the inline extensions/apps install UI for Cocoa.
This required a new .xib (ExtensionInstallPromptInline) that shows store data
(star rating, number of reviews, number of users). Current .xibs
(ExtensionInstallPrompt and ExtensionInstallPromptNoWarnings) were rearranged
slightly to have the icon on the right and use a bulleted list for the list
of permissions (instead of a box) per Cole's mock.

This is the Cocoa side of r99407 (views) and r99205 (GTK).

BUG=93380
TEST=no
Review URL: http://codereview.chromium.org/7835039

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@99695 4ff67af0-8c30-449e-8e8b-ad334ec8d88c