[Sync] Don't attempt to show the configure screen if the backend isn't ready.
We simply return a null flow, which is the same as if we had done nothing. Next
time the user clicks customize, they will again try to create a new flow, and
if the backend has now initialized, we succeed.

BUG=93127
TEST=manually alter the code to ensure the backend doesn't initialize. Attempt
to customize sync on a client that is already signed in.


Review URL: http://codereview.chromium.org/7851003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@100088 4ff67af0-8c30-449e-8e8b-ad334ec8d88c