Move the ViewMsg_FPS and ViewHostMsg_FPS IPC messages out of content to chrome as these are sent and processed by chrome. These messages are now prefixed with "Chrome"
To handle the ChromeViewMsg_GetFPS message in the ChromeRenderViewObserver object we need to expose the value
of the filtered_time_per_frame member in the RenderWidget class. Added a getter to achieve that.

Continuing changes to ensure that IPCs don't span across content and chrome.

BUG=87335
Review URL: http://codereview.chromium.org/7754004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@100178 4ff67af0-8c30-449e-8e8b-ad334ec8d88c