Move the UTF8 text out of .cc files and put them into regular_expressions.txt.
This is temporary.  I think having them in a separate file is too much
maintenance overhead.  So I plan on following up with extending grit to handle
this.  I want to go ahead and get this in to fix people's build and
so the encoding change for pak files can go in.

Review URL: http://codereview.chromium.org/7856001

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@100213 4ff67af0-8c30-449e-8e8b-ad334ec8d88c