Fix memory leak in sync notifier unit tests
When XmppConnection is deleted, its TaskPump member is not deleted immediately.
The destructor equeues a task to the message loop to have it deleted later.  A
comment in ~XmppConnection that explains why it does this.

In the unit tests, the TaskPump would get leaked because the message loop would
never get a chance to run that deletion task.  This commit fixes the problem
by running all pending tasks on that message loop before we exit the test.

BUG=80238,79651
TEST=InvalidationNotifierTest


Review URL: http://codereview.chromium.org/7863005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@100542 4ff67af0-8c30-449e-8e8b-ad334ec8d88c