sync: ensure sync init path doesn't block on CheckTime
The call to RequestEarlyExit (which calls Abort) only happens if the SyncBackendHost has received the initialization callback from the SyncManager.  But during init, the SyncManager could make a call to CheckTime, meaning that call would never be aborted.  This patch makes sure to cover that case.

BUG=93829
TEST=None at the moment :(


Review URL: http://codereview.chromium.org/7862011

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@100543 4ff67af0-8c30-449e-8e8b-ad334ec8d88c