Redirect chrome://extensions to the new chrome://settings/chromeExtensions.
The two modifications to the tests were needed because they were trying to
open chrome://extensions in incognito.

In one case, the test just needed any page (so I switched to about:blank)
and in the other we are testing for a condition that cannot happen anymore
because chrome://extensions could be loaded in incognito but chrome://settings
doesn't allow that -- it shunts the request to the non-incognito profile.

The test was testing that we don't crash in incognito, so I removed that test.

BUG=87377
TEST=Well... type in chrome://extensions and notice it redirects.
Review URL: http://codereview.chromium.org/7888010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@101046 4ff67af0-8c30-449e-8e8b-ad334ec8d88c