Improve fast-shutdown path to prevent flakiness
The current pattern used in the fast shutdown path
is to send a message to the file thread, and assume
that when it responds, then all prior file thread tasks
are done.  The "response" is a quit message.  It is
possible that some other quit messages are pending,
and may cause us to terminate before the "real"
response comes.  This changes carefully waits for the
real response, and pushes back any extra quit messages
that are detected in case other code is waiting for them.

r=darin
BUG=96639
Review URL: http://codereview.chromium.org/7891065

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@101255 4ff67af0-8c30-449e-8e8b-ad334ec8d88c