Make panel titlebars more tab-like on Mac.
When using the default theme, we try to match the default titlebar look-and-feel. When using custom themes, we use the selected/unselected tab resources instead. The .xib update is to change the panel window to be a textured window, so we can retrieve the NSColor to use to paint the titlebar when using the default Chrome theme.

BUG=none
TEST=Open panels and see what they look like.

Review URL: http://codereview.chromium.org/7875011

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@101817 4ff67af0-8c30-449e-8e8b-ad334ec8d88c