Test the impact on stability of not uplodaing any events
Current logs uploaded are swollen.  Test hypothesis that the
cause is extra long lists of "events" by not sending any events.
This will mean we will get stability data, but not event data.

The patch employs a code path that is not used *that* often, so
it has less testing.

r=rtenneti
BUG=97181
Review URL: http://codereview.chromium.org/7971007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@101894 4ff67af0-8c30-449e-8e8b-ad334ec8d88c