PrintPreview: Send a print request only after the print data is ready.
(1) When we print to pdf, we need to wait till the print ready
metafile is available.
(2) When we print a pdf file, we need to wait till the print ready
metafile is available.
(3) When we print to a physical printer, we need to wait till all the
required draft pages are available and loaded in the preview tab plugin.

BUG=94358
TEST=please refer to bug description.


Review URL: http://codereview.chromium.org/7891027

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@101900 4ff67af0-8c30-449e-8e8b-ad334ec8d88c