After a successful ping of the RLZ server, make sure to update the cached RLZ strings for the omnibox and home page access points instead of just invalidating, so that the first time the code tries to use the strings after the ping they are valid.
BUG=None
TEST=Install chrome.  Chrome will start automatically, do not stop it.
Wait for the RLZ ping to be sent.  Perform a search with the omnibox
and make sure an rlz= CGI parameter is present.  Visit the home page and
make the request contains an X-Rlz-String HTTP header.
To perform these tests, install an official chrome build with a non-organic
brand code.

Review URL: http://codereview.chromium.org/7980013

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@102321 4ff67af0-8c30-449e-8e8b-ad334ec8d88c