Convert SHA256_LENGTH from a constant-in-anonymous-enum to a static const.  This defines the constant where it's declared to preserve the existing readability.
Normally this makes things like DCHECK_EQ() unhappy, but when I'd originally tested this I didn't seem to need to make any changes due to that.  Will be watching the trybots...

The original motiviation for this change was to find a way to eliminate some cases of passing anonymous-typed values as template arguments (which happens when you use a value from the enum in e.g. EXPECT_EQ()), which is technically illegal in C++03, though we don't warn about it. Simply naming the enum would have done this, but in general naming enums used to declare constants like this is bizarre ("enum Constants { ... }"?).

BUG=92247
TEST=Compiles
Review URL: http://codereview.chromium.org/7823004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@102369 4ff67af0-8c30-449e-8e8b-ad334ec8d88c