CrOS: Add Discard Tab command link to about:discards
Work in process on discarding a tab (for low memory conditions on CrOS).  Also clean up the strings we use to describe system memory, and eliminate memory as an input for "interestingness" of tabs, since the OOM killer already includes memory as an input.

BUG=none
TEST=Open "about:discards".  Click the "Discard Tab Now" link - the tab at the bottom of the list should close.

Review URL: http://codereview.chromium.org/7983042

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@102397 4ff67af0-8c30-449e-8e8b-ad334ec8d88c