Deflake PrerenderBrowserTest.PrerenderClientRedirectInIframe.
What seems to be happening is that the client redirect can
result in starting and stopping the page loading again.  As
a result, we quit the message loop when the second load
finishes as well as the first, which can occur either while
we're waiting for the result of a Javascript function, or
while we're waiting for the navigation to occur.  Either
case will result in a test failure.

BUG=88973
TEST=PrerenderBrowserTest.PrerenderClientRedirectInIframe (And probably PrerenderBrowserTest.PrerenderClientRedirectToHttpsInIframe)

Review URL: http://codereview.chromium.org/7981033

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@102431 4ff67af0-8c30-449e-8e8b-ad334ec8d88c