Convert PyAuto hooks to multiprofile.
Specific changes:
Add a method to create a new profile to TestingAutomationProvider.
Add a method to open a browser window with a given Profile.
Fix methods in TestingAutomationProvider to use the profile from a browser, instead of the instance variable profile_.
Add a test in ntp.py to make sure ntp pages don't leak between profiles, to test the new methods.

The general idea here is that one can still use profile_ for most tests, but one has the option of creating multiple profiles and opening windows with a specific profile to test multiprofile features.

BUG=87470
TEST=new ntp.py test in CL.

Review URL: http://codereview.chromium.org/7865026

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@102470 4ff67af0-8c30-449e-8e8b-ad334ec8d88c