Provide implementations of NPChannelBase to allow script to use NPAPI objects in the browser process
This functionality is required to implement the Java Bridge on Android.
JavaBridgeChannel is used with NPObjectProxy in the renderer and
JavaBridgeChannelHost is used with NPObjectStub in the browser.

Note that this requires a minor change to NPChannelBase to allow JavaBridgeChannelHost to finish setting up the channel handle.

BUG=96703

Review URL: http://codereview.chromium.org/7906005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@102842 4ff67af0-8c30-449e-8e8b-ad334ec8d88c