Make calls to registerProtocolHandler() update already registered handlers.
This change makes it so that if a site calls rPH() with a template URL that has the same origin as an already registered handler (for the same protocol) then the handler is silently replaced. This avoids confusing infobars for sites that change the URL or title of their protocol handlers - "Would you like to use $1 to handle webcal: links instead of $1?"

This has the effect of making it so that only one protocol handler can be registered for each [protocol, origin] pair. This is desirable at least in the case described above because it means that there will be no stale protocol handlers hanging around in the ProtocolHandlerRegistry that are not intended to be used.

BUG=95158
TEST=Unit tests provided


Review URL: http://codereview.chromium.org/7978018

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@103086 4ff67af0-8c30-449e-8e8b-ad334ec8d88c