Experimentally strictify third-party cookie blocking.
This CL turns on strict third-party cookie blocking by forcing the inclusion of
the `kBlockReadingThirdPartyCookies` command-line flag. The goal is to determine
the effect of a change in cookie-blocking behavior; if we Break The Web™ for
Canary users, we'll drop it. If not, we'll re-evaluate the default behavior of
the `kBlockThirdPartyCookies` pref.

I intend to revert this change after a few Canary cycles.

BUG=98241
TEST=Start Chrome. Is the command-line flag set?


Review URL: http://codereview.chromium.org/8060030

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@103112 4ff67af0-8c30-449e-8e8b-ad334ec8d88c