Add cases to switch in ModelTypeToString
Some sync integration tests will attempt to get the string for the model type
TOP_LEVEL_FOLDER when the tests are running in very verbose mode (--v=2).
Because there was no case for this type in the switch statement, the tests would
run into a NOTREACHED() assertion.

This change adds the case statement necessary to avoid the NOTREACHED in these
tests.

BUG=98082
TEST=MigrationSingleClientTest.AllTypesWithNigoriIndividually


Review URL: http://codereview.chromium.org/8041036

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@103184 4ff67af0-8c30-449e-8e8b-ad334ec8d88c