Rewrites renderer accessibility to not use WebAccessibilityCache.
This fixes the primary underlying cause of a number of bugs where the
browser's accessibility tree got out of sync with the renderer's accessibility
tree, which makes the page seem not functional anymore.

Splits renderer accessibility code out of render_view.cc into its own file.
All code now uses the axID directly from WebCore::AccessibilityObject rather
than WebAccessibilityCache, which can now be deleted. One implication of this
is that the top-level node can now change its ID, so we introduce a new
role ROLE_ROOT_WEB_AREA to distinguish it from an iframe's ROLE_WEB_AREA node.

To solve the problem of getting out of sync, the renderer now
maintains a tree of axIDs that reflects the state of the tree as of
the most recent message sent to the browser. Whenever a new
accessibility notification is posted, it refers to that tree to
determine precisely what information needs to be sent to the browser,
eliminating problems where the renderer was posting notifications
about nodes that the browser didn't know about.

BUG=93095,93232,92716,90787,90768
TEST=updated lots of tests

Review URL: http://codereview.chromium.org/7966013

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@103249 4ff67af0-8c30-449e-8e8b-ad334ec8d88c