[mac] Enable floating bookmarks bar for NTP4 for M15 beta branch.
Details: It was recently decided to disable NTP4's support for bookmarks on the M15 branch.  This was done in the following patch: http://codereview.chromium.org/8043004/

The above CL was committed, merged, and then reverted.  Unfortunately the above doesn't re-enabled floating bookmarks on Mac, so this patch exists to do that.  We will repeat the process.... commit, merge to m15, revert.

BUG=98572
TEST=Verify floating bookmarks displays on NTP4 under M15 branch.
Review URL: http://codereview.chromium.org/8085005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@103306 4ff67af0-8c30-449e-8e8b-ad334ec8d88c