Add a content_shell specific DownloadManagerDelegate.
Also fix a race condition that existed in resource dispatching but was never tickled by Chrome. The problem is if in BufferedResourceHandler::OnReadCompleted(), CompleteResponseStarted calls the next handler's OnResponseStarted. If that pauses the request, then the next handler's OnReadCompleted() would still get called which is not correct.

BUG=90445
Review URL: http://codereview.chromium.org/8065019

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@103331 4ff67af0-8c30-449e-8e8b-ad334ec8d88c