Clean up WebUI task manager CSS and disable selection.
This disables user selections and uses the default cursor in
the task manager dialog; previously, the title and the
headings were selectable (sadly, the possibly-useful-to-copy
data being displayed in the dialog was already
unselectable).

I'm also sorting the CSS rules as instructed by the style
guide.

BUG=chromium-os:20738
TEST=manual

Review URL: http://codereview.chromium.org/8089011

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@103465 4ff67af0-8c30-449e-8e8b-ad334ec8d88c