chromeos: Remove border around dialogs and use RGB.
This gets rid of the one-pixel-wide, dark gray border around
the not-really-a-bubble-anymore BubbleWindow frame used for
Chrome OS dialogs.  It also makes the windows be RGB instead
of RGBA to let the window manager know that it's safe to
draw a shadow around them.

BUG=chromium-os:20997
TEST=manual: checked that dialogs get shadows (in conjunction with http://gerrit.chromium.org/gerrit/8538)

Review URL: http://codereview.chromium.org/8085019

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@103557 4ff67af0-8c30-449e-8e8b-ad334ec8d88c