Fix the problem that sometimes panel on Windows is not set to top-most.
This could happen when we try to make the window top-most after it has been
created without activation. This seems to be a bug or feature on Windows that
some users have been complaining about.

To fix this problem, I changed BrowserFrame::InitBrowserFrame to pass the
keep_on_top flag when the panel is to be created.

I have also added windows specific testing code to check the styles in the
browser test.

BUG=none
TEST=browser test
Review URL: http://codereview.chromium.org/8098007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@103602 4ff67af0-8c30-449e-8e8b-ad334ec8d88c