Don't destroy instant loaders if you don't have to.
As long as the omnibox retains focus, don't delete any instant loaders. This
includes cases where the user is not typing actively, or when the user switches
tabs (but retaining focus in the omnibox). The rationale is that the user may
type at any time, so it's useful to keep the instant loaders ready to process
the user query.

BUG=none
TEST=PreloadInstantTest.*


Review URL: http://codereview.chromium.org/8085010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@103938 4ff67af0-8c30-449e-8e8b-ad334ec8d88c