Show sync promo at startup
This CL changes the startup behavior so that for new Chrome installs we now show the chrome://syncpromo page instead of the new tab page.

We also do the same thing when a user creates a new profile.

This CL also keeps track of the number of times the promo page has been shown so that we only show it 10 times.

In a later CL I'll add code to suppress the promo if the user is already signed into sync or if they have clicked the "Skip setup for now" button.

BUG=97779
TEST=


Review URL: http://codereview.chromium.org/8093016

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@104054 4ff67af0-8c30-449e-8e8b-ad334ec8d88c