chromeos: Start Chrome with a non-maximized window on Linux desktop by default.
Previously developers needed to specify --chromeos-frame to start
Chrome with a non-maximized window on Linux desktop.

This patch removes --chromeos-frame and change the default behavior
to start with a non-maximized window on Linux desktop. This will make
it easier to develop on Linux develop. One less thing to remember is
a good thing.

Don't panic! You can still start Chrome with a maximized window with
--start-maximized.


BUG=chromium-os:21346
TEST=confirm that Chrome starts with a non-maximized window on Linux desktop, and in a maximized window with --start-maximized. On Chrome OS, Chrome starts with a maximized window.


Review URL: http://codereview.chromium.org/8133017

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@104306 4ff67af0-8c30-449e-8e8b-ad334ec8d88c