Add "data synced" concept to sync integration test
The documentation for AwaitSyncCycleCompletion claimed it could be used
to wait for a single sync cycle.  In reality, it would wait for many
cycles until snap->unsynced_handles == 0.  Although you'd never guess
this by looking at it, this condition implies that the client has done
at least one update since it committed items to the server.  This means
that its progress marker is likely to be up to date, which is a very
useful property.  See the comments in crbug.com/48989 for more info.

When issue 94670 is fixed it will no longer be guaranteed that a commit
will be followed by an update.  So the snap->unsynced_handles == 0
condition might never evaluate to true.  However, even though a client's
progress marker is stale, it might have its data be fully up to date and
committed to the server.  That's the state that the newly introduced
WaitForDataSynced function looks for.

This commit renames AwaitSyncCycleCompletion to AwaitFullSyncCompletion,
which more accurately reflects what the function actually does.

This commit should not change the integration test behaviour in any way.
The new AwaitDataSyncCompletion and related functions are not used.  The
AwaitFullSyncCompletion should behave exactly the same as the old
AwaitSyncCycleCompletion did.

BUG=94990
TEST=sync_integration_tests


Review URL: http://codereview.chromium.org/8139015

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@104343 4ff67af0-8c30-449e-8e8b-ad334ec8d88c