Properly handle vertically flipping the texture.
Our old mechanism for drawing external textures from the GPU process that are vertically flipped does not work if the corresponding layer has a hole in it (we don't want to flip the hole). This patch handles the flip at the texture map where it belongs.

BUG=none
TEST=with --use-gl=egl on a debug build of TOUCH_UI, go to a page with a combo box (such as news.google.com); open the combo box; you should not see a bright purple hole.

Review URL: http://codereview.chromium.org/8177007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@104348 4ff67af0-8c30-449e-8e8b-ad334ec8d88c