Sync Promo: Open NTP if sync promo is disabled
If navigation to the sync promo page was disabled while a navigation was pending it could cause a CHECK() in TabContents::NavigateToEntry to fail.

The problem was that GetWebUIFactoryFunction() would return NULL if the promo was disabled. To prevent this kind of problem I'm simply returning the NTP WebUI instead.

BUG=99359
TEST=


Review URL: http://codereview.chromium.org/8176015

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@104394 4ff67af0-8c30-449e-8e8b-ad334ec8d88c