profile: Add a way for each ProfileKeyedServiceFactory to specify its user prefernces.
To allow us to compile individual features in/out of a chrome build, we need to
attack browser_prefs.cc, which is a giant static registration. This used to be
done through another big code path starting in ProfileImpl which call
browser::RegisterUserPrefs() which linked to other static methods. Now this is
done through overridding ProfileKeyedServiceFactory::RegisterUserPrefs(), and
there are no additional dependencies.

BUG=none
TEST=none


Review URL: http://codereview.chromium.org/7901027

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@104412 4ff67af0-8c30-449e-8e8b-ad334ec8d88c