Sync Promo: Skip the configure pane when possible
To simplify the first run experience we now skip the configure pane if no passphrase is set.

To test this I logged in with a sync account with no passphrase set and verified that the configure pane was skipped. I then logged in with a sync account with a passphrase set and verified that the configure pane was not skipped.

BUG=98844
TEST=


Review URL: http://codereview.chromium.org/8220029

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@104858 4ff67af0-8c30-449e-8e8b-ad334ec8d88c