Assign urls to RenderProcessHosts by if they refer to isolated apps, or use the extension scheme.
With extensions, a renderer process may be is granted privileges that are different from normal web renderers.  This defines two coarse privilege buckets: one for extensions, and the other for extensions that have isolated storage.  Only sites that belong to the same privilege bucket will be assigned to the same RenderProcessHosts.

BUG=60844
TEST=new browser_tests

Review URL: http://codereview.chromium.org/8186003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@105029 4ff67af0-8c30-449e-8e8b-ad334ec8d88c