Move content_plugin_client.h to content\public\plugin.
The plugin_carbon_interpose_mac.cc which intercepts system API's in plugin processes on the Mac has been moved to

content\plugin. The plugin_carbon_interpose shared library target which lives in chrome.gyp refers to this file.

This is a short term workaround while we look into switching the interposing mechanism on the mac.

Changes as part of creating a Content API

BUG=98716
Review URL: http://codereview.chromium.org/8224021

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@105128 4ff67af0-8c30-449e-8e8b-ad334ec8d88c