Replace bool with void, where return values are not used in the callers.
This is a preparation work for moving login manager D-Bus code from libcros
to Chrome. We want to make semantics clearler beforehand.

Besides, making them void makes the migration easier. If we want to keep
the boolean return values, we either need to 1) call methods synchronously
(this is what we do currently, which is not desirable), 2) return methods
asynchronously. If needed, we should go with 2), but for now making them
void makes things any worse, as the return values aren't used now.

TEST=chrome builds
BUG=chromium-os:16555

Review URL: http://codereview.chromium.org/8226012

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@105263 4ff67af0-8c30-449e-8e8b-ad334ec8d88c