[Sync] Remove unnecessary DCHECK in pref model associator.
We assumed that when a preference becomes unmanaged, it will be user
controlled as well. This isn't the case if it was a default value previously.
This gets handled properly in the association code though, so just remove the
DCHECK and make the comment clearer.

BUG=97611
TEST=unit_tests --gtest_filter="*ProfileSyncServicePreference*"


Review URL: http://codereview.chromium.org/8266001

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@105379 4ff67af0-8c30-449e-8e8b-ad334ec8d88c