Fixes bug in instant where we would end up incorrectly using the preview when we shouldn't. Here's the sequence that would trigger it:
1. focus the omnibox (which triggers loading the InstantLoader).
2. Type in a string that'll autocomplete to a url.
3. Arrow over a search suggestion to a non-search entry.
4. Press enter.

When you arrow over a non-search we'll hide (what was
DestroyPreviewContentsAndLeaveActive) the preview. But if between
steps 3 and 4 we get a response back from the page with suggestions
we'll set displayable_ to true and think everything is up to
date. This leads to IsCurrent returning true when it isn't.

To fix this I've nuked is_active(), which was a bit confusing anyway
and added is_out_date_ (still confusing, but at least it's private).

BUG=100368
TEST=covered by test, see bug for test scenario.
R=sreeram@chromium.org,ben@chromium.org
TBR=ben@chromium.org

Review URL: http://codereview.chromium.org/8298005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@105664 4ff67af0-8c30-449e-8e8b-ad334ec8d88c