Crash fix.
We never reset |tab_contents_| to NULL anywhere (and indeed we can't,
even in ReleasePreviewContents(), because browser.cc accesses it). We
were protected from acccessing a dangling tab_contents_ before, due to
the is_active() check that preceded a call to PrepareForCommit(). That
was removed in http://crrev.com/105664 however; hence this fix.

BUG=100521
TEST=none

Review URL: http://codereview.chromium.org/8329020

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@106075 4ff67af0-8c30-449e-8e8b-ad334ec8d88c