Add asyncBlocking support to webRequest event listeners.
If an event listener is registered with "asyncBlocking" listed in the extraInfoSpec, then a callback is passed into the listener in addition to the details argument. The callback expects a BlockingResponse argument, and MUST be invoked by the listener at some point so the request can proceed.

This is supported for the onAuthRequired listener only at this point.

BUG=32056
TEST=Write an extension with an onAuthRequired listener set with "asyncBlocking" in the extra parameters, validate that it works.


Review URL: http://codereview.chromium.org/8354014

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@106543 4ff67af0-8c30-449e-8e8b-ad334ec8d88c