[filebrowser] Never display large images in thumbnails and previews
When there is no thumbnail encoded in an image file we use
the whole image for preview/thumbnail. For large files this kills
performans. This patch limits the file size to 1Mb, for larger files
a placeholder icon is displayed.

Also some more fixes.

BUG=chromium-os:21853,chromium-os:21066
TEST=

Review URL: http://codereview.chromium.org/8353035

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@106703 4ff67af0-8c30-449e-8e8b-ad334ec8d88c