Add a new field trial for Instant.
In this mode, we issue queries as the user types, but don't show the preview.
The name HIDDEN seems more appropriate here, so I've usurped that name for this
trial, renaming the earlier one to SILENT (since it doesn't speak the queries
over the wire).

BUG=none
TEST=interactive_ui_tests InstantFieldTrial*

Review URL: http://codereview.chromium.org/8370020

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@107053 4ff67af0-8c30-449e-8e8b-ad334ec8d88c