Prevent extension and bookmark popups from interfering with close window/tab shortcuts.
When a balloon popup is open, the target of the close window/tab keyboard shortcut is still the main browser window.  Update our custom menu item rejiggering code to pay attention to the main window, rather than to the popup window, when a popup is open.

BUG=95169
TEST=(see bug)


Review URL: http://codereview.chromium.org/8302021

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@107666 4ff67af0-8c30-449e-8e8b-ad334ec8d88c