Relanding this as the earlier attempt broke chrome frame tests.
Register the RenderViewHost instance created when a desktop notification bubble is displayed in CF
with the AutomationResourceMessagefilter. This ensures that any HTTP requests issued by this bubble
go through the host network stack implemenation. I also fixed a bug in the pending http jobs resumption
code in the AutomationResourceMessagefilter which could cause a job to be registered in the old filter
and the new filter for a while.

Fixes bug http://code.google.com/p/chromium/issues/detail?id=64251

BUG=64251
Review URL: http://codereview.chromium.org/8414030

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@107795 4ff67af0-8c30-449e-8e8b-ad334ec8d88c