aura: Work around RenderWidgetHostTest.Background segfault.
This test was creating an aura::Window but not calling its
Init() method, resulting in a segfault when
RenderWidgetHostViewAura::SetBackground() tried to access
the window's layer.

I have a more ambitious change in progress to add a
RenderWidgetHostView::InitAsChild() method and update a
bunch of callers; this is just a quick workaround so try
bots will stop failing.

BUG=102450
TEST=manual: test no longer segfaults

Review URL: http://codereview.chromium.org/8432020

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@108048 4ff67af0-8c30-449e-8e8b-ad334ec8d88c