Refactored TypedUrlChangeProcessor to treat ADD and UPDATE similarly.
Updated TypedUrlChangeProcessor to handle ADDs and UPDATEs in the sync DB identically.

Also removed our code that explicitly sets the title in the history DB as that
should not be necessary (it's already set by UpdateURLRow).

Finally, removed our code that was setting last_visit_time since that is also
updated automatically by the history code when we add/remove visits. We still
set it when we initially add the URL to the DB, but that's only because the
history code requires it - we technically don't need to since we always manually
add visits immediately afterward.

BUG=101633
TEST=existing tests suffice


Review URL: http://codereview.chromium.org/8414043

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@108188 4ff67af0-8c30-449e-8e8b-ad334ec8d88c