[Mac] Fix crash for "settings" case of protector bubble.
If the current page was not NTP, the code to go to settings caused a
new tab to be brought up, which caused the bubble to be closed (by
base bubble controller), and then the second call to -close caused a
crash.  The duplicate calls to -close should be fine, but the message
to a freed |error_| needed to be dropped.

BUG=102681
TEST=Arrange to be on a website (like google) at startup.  Arrange to get the protector bubble up on startup, and click the settings option.  Should go to the settings, and not crash.


Review URL: http://codereview.chromium.org/8372088

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@108420 4ff67af0-8c30-449e-8e8b-ad334ec8d88c