aura: Avoid pref crash when using WebKit compositor.
This fixes an issue where we'd attempt to read
auth.allow_cross_origin_prompt before registering it when
building with use_webkit_compositor=1.

We were formerly registering the pref when the IO thread was
created.  It looks to me like it only gets accessed in
BrowserProcessImpl, so let's register it there instead.

BUG=103314
TEST=manual: checked that we no longer crash after building with use_webkit_compositor=1

Review URL: http://codereview.chromium.org/8500002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@109045 4ff67af0-8c30-449e-8e8b-ad334ec8d88c