Make it possible to include another file and port print_preview unit test to unit_tests.
Print Preview unit_tests were good to tackle because it brought in a dependency on the need
to include another js file.

This also hilighted another need - that of being able to include non-generative js files in gyp
and have them copied to the test_data dir, while doing both generation and copy for gtest files.
I changed the "extension" of files which generate C++ with gtest-like syntax to .gtestjs and
added a copyjs rule to the chrome_tests.gypi.

FWIW, I believe this is mostly needed for unit_tests and only applied it there, although I would
be amenable to also making this change for the webui tests to make it clear that they are not
plain js files.

R=dpapad@chromium.org
BUG=101443,102222
TEST=unit_tests --gtest_filter=PrintPreviewUtilsUnitTest.*

Review URL: http://codereview.chromium.org/8438063

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@109310 4ff67af0-8c30-449e-8e8b-ad334ec8d88c