Include PPAPI plugin crashes in the plugin crash data.
This should enable us to see PPAPI plugin crashes in the plugin crashes column
on the stability dashboard. Note that each plugin (PPAPI or regular) is included
separately in the metrics log and aggregated at the server side. This allows us
to slice and dice the data later, if necessary.

Also:

- Remove a couple of obsolete FRIEND_TEST declarations.

- Use base file name as plugin name when the name of the PPAPI plugin to be
  launched is empty. This is mostly so that the OutOfProcessPPAPITests can pass
  but also matches the logic in PepperPluginInfo::ToWebPluginInfo().

BUG=chromium-os:22465
TEST=git try, unit tests, kill -SEGV ppapi process, check that it's counted in
the plugin prefs and transmitted in the metrics log:
<pluginstability filename="nfGm7uk2q0xBpC3X5CA95A==" launchcount="3"
                 instancecount="1" crashcount="2"/>


Review URL: http://codereview.chromium.org/8501005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@109627 4ff67af0-8c30-449e-8e8b-ad334ec8d88c