Remove GPU blacklist auto-update.
Also, use the software_rendering_list.json in deps, so it can be shared by all branches and main trunk.

Note that this CL is the minimum because we need to merge back to M16.  The remaining cleanup, including removing the dead code, simplifying GpuDataManager, refactoring, will be done in a separate CL.

BUG=103851
TEST=try bots
Review URL: http://codereview.chromium.org/8511067

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@109738 4ff67af0-8c30-449e-8e8b-ad334ec8d88c