deprecate webui.css
This file is really only useful to downloads and history. All other includes of the file are senseless copy-pasta as far as I can tell. There is also the problem that due to its poor name, it became a dumping ground for generic stuff like displaytable, and then was included into chrome://settings because of it. Move out the bits that should live elsewhere, remove the rules that are not used anywhere, rename the file to head off re-use.

Also move gpu_internals stuff into gpu_internals/, and remove completely duplicated file webui2.css.

BUG=103688
TEST=manually checking gpu internals, media internals, policy, tracing, options, history, downloads

Review URL: http://codereview.chromium.org/8511046

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@109765 4ff67af0-8c30-449e-8e8b-ad334ec8d88c