content: Split off ChromeTabContentsViewGtk from TabContentsViewGtk. This new class contains almost all chrome specific details from TabContentsViewGtk, leaving TabContentsViewGtk as a pure content/ class (soon to be moved out of chrome/).
Why subclass instead of composing? First, because the X vs ChromeX is fairly
common already, but mostly because large chunks of the gtk ui code assume
there's a single widget that can be switched in and out on tab showing and that
this is TabContentsViewGtk::GetNativeView(). (This also means it is easy to
address the other join points, like sad tabs or menus.)

BUG=none
TEST=none


Review URL: http://codereview.chromium.org/8344061

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@110236 4ff67af0-8c30-449e-8e8b-ad334ec8d88c