Add ChromeBrowserMainExtraParts for non main parts.
This was originally submitted as http://codereview.chromium.org/8539038/ which was reverted because of a failing test on win.

This reverts the code back to a single instance of BrowserMainParts, with auxillary parts (Gtk, Views, Aura, Touch) implemented from a new base class, ChromeBrowserParts, which has a Chrome specific interface, allowing initialization to be better subdivided.

This should fix the notifications auto tests.

This CL also addresses the test failure on win (and a test failure on linux_chromeos triggered by the windows fix).

BUG=103821
TEST=Make sure all tests and autotests run

TBR=jam@chromium.org

Review URL: http://codereview.chromium.org/8598005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@110710 4ff67af0-8c30-449e-8e8b-ad334ec8d88c