Move the ProcessWatcher methods out of content/common/process_watcher into base/process_util, alongside the other process methods.
The only non-trivial move change is to the Windows implementation, where I changed KillProcess to use an exit code of kProcessKilledExitCode instead of content::RESULT_CODE_HUNG. cpu said that the existing code was incorrect, since GetTerminationStatus() should be mapping that result to TERMINATION_STATUS_PROCESS_WAS_KILLED. So I changed the exit code to kProcessKilledExitCode. This might make the UMA stats for killed processes to go up (and crashed to go down), but that will be an accounting change and should be zero-sum.

BUG=98716
Review URL: http://codereview.chromium.org/8674003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@111371 4ff67af0-8c30-449e-8e8b-ad334ec8d88c