Move GPU preliminary info collection back to UI thread.
Apparently doing it on FILE thread causes a race, which causes crbug 83153.

In order to move back to UI thread, we have to drop chrome channel support in GPU blacklist (it's a FILE operation); also, we have to move the collection of ATI driver version in linux to GPU process.

Note that this is the minimum change required for the move.  Further cleanup will follow once this is merged safely back to M16.

BUG=97535
TEST=unittest
Review URL: http://codereview.chromium.org/8670004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@111390 4ff67af0-8c30-449e-8e8b-ad334ec8d88c