Change sync histogram logging from Boolean to Enumeration
Histogram logging for custom selection of sync datatypes is currently
done using booleans. It's better to use an enumeration, since the
datatypes are related, and it makes sense to have a unified view of all
sync datatypes.

This patch adds a new enum called UserSelectableSyncTypes, and changes
the UMA histogram logging from booleans to an enumeration.

BUG=96507
TEST=Customize sync in different ways and examine about:histograms.

Review URL: http://codereview.chromium.org/8680022

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@111771 4ff67af0-8c30-449e-8e8b-ad334ec8d88c