TabRestoreService improvements
this patch brings more code from the views into the model (TabRestoreService) so the views don't all duplicate each other, or forget to duplicate each other's behavior and then act inexplicably different. For example, only RecentlyClosedHandler was removing duplicates, and it doesn't make sense to diverge on that point. (In the end, this patch actually removes all de-duping because two tabs may diverge in their history even if their current navigation is the same.)

This also fixes the issue where filtering was done after the list had been pruned to 10 entries. The issue with doing this is that filtered entries (such as New Tabs or duplicate tabs) counted against the total of 10 entries, but wouldn't show up in the list, so you'd frequently show <10 tabs even though there are 10 valid, closed tabs we could show.

BUG=83211, 81805
TEST=
1. open 10 tabs, close them all. Start opening and closing new tabs --- the recently closed list shouldn't be affected.
2. The jumplist on Windows and the global history menu on mac/gtk should not show duplicate entries
3. if you close a window with 2 normal tabs and one ntp in it, it should show up as "3 tabs", not "2 tabs"

Review URL: http://codereview.chromium.org/8603008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@111885 4ff67af0-8c30-449e-8e8b-ad334ec8d88c