Controlled settings indicator and bubble for chrome://preferences.
The idea is that every setting that is under control by an external entity (e.g. policy, extension) will be annotated with an icon that, when click on, will show a bubble with more information. This implements the indicator icon and bubble. Going through the prefs UI and inserting the thing at the appropriate places is left for subsequent CLs :)

BUG=chromium:104955
TEST=No observable changes.

Review URL: http://codereview.chromium.org/8480037

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@111897 4ff67af0-8c30-449e-8e8b-ad334ec8d88c