Add url and is_main_frame to DidFinishLoad and DidFailLoad.
This will add 2 fields: validated_url and is_main_frame to
the DidFinishLoad TabContentsObserver callback and the
associated view message and will add a DidFinishLoad 
TabContentsObserver callback with the associated view
message.
The change should not have any impact on current behavior.

This is needed for the Chromium port on Android.

BUG=none
TEST=base_unittests,content_unittests,browser_tests


Review URL: http://codereview.chromium.org/8384028

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@111912 4ff67af0-8c30-449e-8e8b-ad334ec8d88c