Tweak GAIA info in ProfileInfoCache
If the user's GAIA name or picture was empty then we would end up displaying a blank name or icon in the multi-profiles UI.

Fix was to fallback to the profile name and icon if the GAIA name and icon were empty.

Another problem was that if no GAIA picture was set then every call to  ProfileInfoCache::GetGAIAPictureOfProfileAtindex() would cause a task to be posted to the FILE thread.

To fix this and make the code more robust I changed setter to save the picture's file name to preferences. This means the getter can now just check the pref and easily know if the file will be on disk or not. This also means that it's easy for us to change the name of the file in the future.

BUG=91241
TEST=

Review URL: http://codereview.chromium.org/8662024

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@112123 4ff67af0-8c30-449e-8e8b-ad334ec8d88c