Replace TOUCH_UI condition in WebUI with dynamic flag.
Touch-specific builds are being retired.  Instead we want to be able to enable touch-friendly UI dynamically.  For now I'm adding a --touch-optimized-ui flag which can be used to enable the touch-optimized CSS tweaks we've made to WebUI.

This undoes the tweak to use the non-ChromeOS language UI for touch_ui builds.  The conditional HTML for the add-language-overlay-page is non-trivial to convert to something dynamic - I'll do something in a separate changelist.

Also fix a number of issues where overrides of WebUI virtuals (like RenderViewCreated) weren't properly delegating back to their base class (or were delegating to WebUI instead of the immediate base ChromeWebUI).

Also removes the Touch-specific User-Agent string.

BUG=105315
TEST=


Review URL: http://codereview.chromium.org/8695007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@112387 4ff67af0-8c30-449e-8e8b-ad334ec8d88c